Rename folder formula to rendering, reorganize nodes into nodes folder
