+++ src/rendering/formula-canvas.tsx
