+++ src/rendering/nodes/node.tsx
