+++ src/rendering/util/label-node.ts
